requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.1
h11==0.14.0
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
playwright==1.40.0
protobuf==4.25.2
pyarrow==14.0.2
pydantic==2.5.3
pydantic-core==2.14.6
pydeck==0.8.1b0
pyee==11.0.1
pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
schedule==1.2.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.32.0.post1
streamlit==1.29.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing-extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
validators==0.22.0
zipp==3.17.0
